Added self-hosting (console) project.
This application still fails. I can start the application and I see no
errors; however, when I attempt to invoke a service, it fails.
Strangely, the console **does not** print the exception and stack trace;
instead, the WcfTestClient application catches this exception and pops
up a window with some details. I can break on the code that throws an
exception (constructing a new ZzaDbContext()) which occurs on the server
side. This exception is captured by the WCF machinery and sent across
the wire to the client (without raising any issue on the server).
(Although it initially seems strange, I believe this is the as-designed
behavior. When designing a service, one wants errors that occur during
requests to propogate back to the client - even if the errors may occur
because the server is "misconfigured." Sigh...