Began trying to test client without UI. Failed.
I made some progress; however, when I ran the unit tests, the unit test
code reported that it could not find the default endpoint. I suspect
this error results from the app.config file being defined in the
Zza.Client project instead of the Zza.Client.Tests.Integration project.

I chose not to pursue this further so that I could finish the videos.